CH14/requirements.txt
python-dotenv>=1.0.1,<2.0.0
langchain-core==0.3.0
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai<3.0.0
langgraph>=0.2.14,<0.3.0
